requirements.txt
PyQt6>=6.5.0
pycryptodome>=3.19.0
markdown>=3.5.0
python-docx>=1.1.0
reportlab>=4.0.0
ebooklib>=0.18.0
httpx>=0.25.0
Pygments>=2.17.0
